Fix wrong behavior with class metadata
The ClassMetadataFactory was broken. By adding a test it
put in evidence that many things doesn't work as expected.

This commit simplifies the code and transform the Iterator into a
Decorator (still doing its job!).

The ugly ClassMetadataDecorator is removed because we can now
manage the instantiation of the ClassMetadata (this was an issue).