Add new feature "Domain rules"
The domain rules are listeners that may do actions on the events dispatched
by the model.

This commit contains a feature that register automatically
a rule that instanciates `DomainRuleInterface`. It's still possible
to add a new domain rule as listener manually by adding a special tag.

This commit also include a page of doc with example to learn more about
this feature.