Add doctrine bundle in dependencies
This is a requirement in order to make the Symfony Recipe work. Of
course it's sad because the library will enforce the installation of the
doctrine bundle. But at this moment I don't think anybody installed this
library without Symfony and the Doctrine bundle. Also the bundle present
in our bridge is dependent of the Symfony Doctrine Bundle.

Taking all of this, I think it's not a bad thing and we will reconsider
if it's a problem some day.

Notice that it also contains doctrine orm requirement, which is not
contained by the doctrine bundle.